0329 금 김태민 휴가 쿼리테스트 완료 및 페이지 연결

+++ GreenLight/src/main/java/com/green/light/vo/EmployeeVo.java
@@ -30,7 +30,7 @@ public class EmployeeVo {
 	private String join_day;
 	private String exit_day;
 	private String etype;
-	private int leave;
+	private float leave;
 	private String auth;
 	private String profile;
 	private int fail;
